Add files via upload
Created server and client for a UDP network. The client has the role of UDP file sender. It reads data from a file and sends it over a UDP connection to a remote receiver. The program uses sockets to establish a connection and transmit data. The server has the role of a UDP file receiver. It receives files over a UDP connection and saves them to the local filesystem.